+++ src/bool/wasm32/bvec4a.rs
@@ -127,6 +127,20 @@ impl BitOrAssign for BVec4A {
     }
 }
 
+impl BitXor for BVec4A {
+    type Output = Self;
+    #[inline]
+    fn bitxor(self, rhs: Self) -> Self {
+        Self(v128_xor(self.0, rhs.0))
+    }
+}
+
+impl BitXorAssign for BVec4A {
+    #[inline]
+    fn bitxor_assign(&mut self, rhs: Self) {
+        *self = self.bitxor(rhs);
+    }
+}
 impl Not for BVec4A {
     type Output = Self;
     #[inline]