+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+* Implemented `BitXor` and `BitXorAssign` traits for `bool` vectors.
+
 ### Fixed
 
 * Fixed some performance regressions in affine and matrix determinant and